requirements.txt
flet
fpdf2
Pillow
pytz
flet-runtime
requests